# BUG: Fix unique handling of nan entries.