--- a/doc/release/upcoming_changes/18070.improvement.rst
+++ b/doc/release/upcoming_changes/18070.improvement.rst
@@ -1,0 +1,6 @@
+``np.unique`` now returns single ``NaN``
+----------------------------------------
+When ``np.unique`` operated on an array with multiple ``NaN`` entries,
+its return included a ``NaN`` for each entry that was ``NaN`` in the original array.
+This is now improved such that the returned array contains just one ``NaN`` as the
+last element.