--- a/doc/release/upcoming_changes/18070.improvement.rst
+++ b/doc/release/upcoming_changes/18070.improvement.rst
@@ -4,3 +4,9 @@ When ``np.unique`` operated on an array with multiple ``NaN`` entries,
 its return included a ``NaN`` for each entry that was ``NaN`` in the original array.
 This is now improved such that the returned array contains just one ``NaN`` as the
 last element.
+
+Also for complex arrays all ``NaN`` values are considered equivalent
+(no matter whether the ``NaN`` is in the real or imaginary part). As the
+representant for the returned array the smallest one in the
+lexicographical order is chosen - see ``np.sort`` for how the lexicographical
+order is defined for complex arrays.