--- a/numpy/lib/tests/test_arraysetops.py
+++ b/numpy/lib/tests/test_arraysetops.py
@@ -533,7 +533,7 @@ class TestUnique:
 
         # test for ticket 2111
         a = [2, np.nan, 1, np.nan]
-        assert_array_equal(np.unique(a), [1, 2, np.nan, np.nan])
+        assert_array_equal(np.unique(a), [1, 2, np.nan])
 
     def test_unique_axis_errors(self):
         assert_raises(TypeError, self._run_axis_tests, object)