BUG: Unique and nan entries (closes #2111)

When unique operates on an array with multiple NaN entries its return includes a NaN for each entry that was NaN in the original array.

This is my first PR in an open-source project, so please apology any mistakes. In this fix, I used the code suggested by [ufmayer](https://github.com/numpy/numpy/issues/2111#issuecomment-639957194). It works correctly. I have added unit tests as well.

The performance impact is negligible as shown by the benchmark I have created.
`python runtests.py --bench bench_lib.Unique`
Before change:
```
[100.00%] ··· ============ ============= ============ ============= ============= =============
              --                                       percent_nans
              ------------ --------------------------------------------------------------------
               array_size        0           0.1           2.0           50.0          90.0
              ============ ============= ============ ============= ============= =============
                  200        7.47±0.2μs   7.58±0.1μs   7.24±0.08μs    7.27±0.2μs    7.14±0.7μs
                 200000     13.2±0.09ms   13.2±0.4ms    13.2±0.3ms   9.63±0.07ms   4.72±0.08ms
              ============ ============= ============ ============= ============= =============
```
After change:
```
[100.00%] ··· ============ ============ ============= ============ ============ =============
              --                                      percent_nans
              ------------ ------------------------------------------------------------------
               array_size       0            0.1          2.0          50.0          90.0
              ============ ============ ============= ============ ============ =============
                  200       9.81±0.1μs    9.85±0.2μs   13.4±0.2μs   13.3±0.2μs     13.2±1μs
                 200000     13.1±0.2ms   13.2±0.06ms   13.3±0.1ms   9.68±0.1ms   4.56±0.04ms
              ============ ============ ============= ============ ============ =============
```