@ftrojan 2020-12-26
I know that `nan = np.full(shape=1, fill_value=np.nan, dtype=aux.dtype)[0]` does not shine, but it does the work and the suggested `np.array(np.nan, dtype=aux.dtype)` was failing on `ValueError: Could not convert object to NumPy datetime`.